% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordSwap.R
\name{recordSwap}
\alias{recordSwap}
\title{Title}
\usage{
recordSwap(data, similar, hierarchy, risk, hid, th, swaprate, seed = 123456L)
}
\arguments{
\item{data}{}

\item{similar}{}

\item{hierarchy}{}

\item{risk}{}

\item{hid}{}

\item{th}{}

\item{swaprate}{}

\item{seed}{}
}
\description{
Title
}
