% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recordSwap_cpp}
\alias{recordSwap_cpp}
\title{Targeted Record Swapping}
\usage{
recordSwap_cpp(
  data,
  similar,
  hierarchy,
  risk,
  hid,
  th,
  swaprate,
  carry_along,
  seed = 123456L
)
}
\arguments{
\item{data}{micro data set containing only integer values. A data.frame or data.table from R needs to be transposed beforehand so that data.size() ~ number of variables - data.[0].size ~ number of records per record.
\strong{NOTE:} \emph{data has to be ordered by hid beforehand.}}

\item{similar}{List where each entry corresponds to column indices of variables in \code{data} which should be considered when swapping households.}

\item{hierarchy}{column indices of variables in \code{data} which refere to the geographic hierarchy in the micro data set. For instance county > municipality > district.}

\item{risk}{column indices of variables in \code{data} which will be considered for estimating the risk.}

\item{hid}{column index in \code{data} which refers to the household identifier.}

\item{th}{integer defining the threshhold of high risk households (k-anonymity). This is used as th <= counts.}

\item{swaprate}{double between 0 and 1 defining the proportion of households which should be swapped, see details for more explanations}

\item{carry_along}{integer vector indicating additional variables to swap besides to hierarchy variables.
These variables do not interfere with the procedure of finding a record to swap with or calculating risk.
This parameter is only used at the end of the procedure when swapping the hierarchies.}

\item{seed}{integer defining the seed for the random number generator, for reproducability.}
}
\value{
Returns data set with swapped records.
}
\description{
Applies targeted record swapping on micro data set, see \code{?recordSwap} for details.
\cr
\strong{NOTE:} This is an internal function called by the R-function \code{recordSwap()}. It's only purpose is to include the C++-function recordSwap() using Rcpp.
}
