% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setLevels_cpp}
\alias{setLevels_cpp}
\title{Define Swap-Levels
Define hierarchy levels over which record needs to be swapped according to risk variables.}
\usage{
setLevels_cpp(data, hierarchy, risk, hid, th)
}
\arguments{
\item{data}{micro data set containing only numeric values.}

\item{hierarchy}{column indices of variables in \code{data} which refere to the geographic hierarchy in the micro data set. For instance county > municipality > district.}

\item{risk}{column indices of variables in \code{data} which will be considered for estimating the risk.}

\item{hid}{column index in \code{data} which refers to the household identifier.}

\item{th}{integer defining the threshhold of high risk households (k-anonymity).}
}
\value{
Integer vector with hierarchy level over which record needs to be swapped with.
}
\details{

}
